Updated Firefly Trigger
Made the firefly trigger spawn the particle system in the players viewcone pointing away from the player. Additionally made the firefly lined particles send particles in one big burst. Lastly just added a variable that sets the max number of particles that the system can throw in 1 huge burst (don't make it go above 100 and if you do change the number of particles being shot in 1 burst to be higher)