Climbing Action (+Animation) (+Input)
Added ClimbAction script which allows the player to climb when appended to PlayerPrefab.
Changes PlayerInputActions and PlayerInput script to allow climbing input.
Adds Climbing State script and adds climb state code to the player state machine script.

NOTE: There is supposed to be a grounded check somewhere in the ClimbAction script that is currently not filled out. This will have to reference some "isGrounded" method for the player.