Created PlayerFallDamage Script
created the player fall damage script which is expected to be added to the player object. This must be setup with either an animation curve or the config for animaiton curve needs to be toggled off. There are issues with the character being too floaty for any real fall damage to take effect because fall damage is based on vertical velocity.